feat: use uv for package management (#77)
<!--
Thank you for contributing to Ruff! To help us out with reviewing,
please consider the following:

- Does this pull request include a summary of the change? (See below.)
- Does this pull request include a descriptive title?
- Does this pull request include references to any relevant issues?
-->

## Summary

`uv` is great. So we should use it where we can. 

The cache had to be removed because the workflow breaks currently.
https://github.com/actions/setup-python/issues/822
When this is solved, caching can be brought back in.

## Test Plan

Ran workflow on my local repo.